% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_O_matrix.R
\name{build_O_matrix}
\alias{build_O_matrix}
\title{Build an orthogonal matrix given a set of real parameters}
\usage{
build_O_matrix(entries)
}
\arguments{
\item{entries}{A vector of entries to build the orthogonal matrix.}
}
\value{
An orthogonal matrix of dimensions according to the given vector
}
\description{
Build an orthogonal matrix given a set of real parameters
}
\examples{
build_O_matrix(0) #Identity
}
