% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcc_cm_bio1.R
\docType{data}
\name{cmcc_cm_bio1}
\alias{cmcc_cm_bio1}
\title{cmcc_cm bio1 time serires}
\format{
A raster time serires with 30 years of bio01 data
}
\usage{
cmcc_cm_bio1
}
\description{
DATASET_DESCRIPTION
}
\details{
The data comes from chelsav1 time serires for
cmcc_cm model. We compute bio01 (Anual tas average) for a 30 years time serires
}
\keyword{datasets}
