% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcc_cm_bio12.R
\docType{data}
\name{cmcc_cm_bio12}
\alias{cmcc_cm_bio12}
\title{cmcc_cm bio12 time series}
\format{
A raster time serires with 30 years of bio01 data
}
\usage{
cmcc_cm_bio12
}
\description{
DATASET_DESCRIPTION
}
\details{
The data comes from chelsav1 time series for cmcc_cm model.
We compute bio12 (Anual precipitation) for a 30 years time series
}
\keyword{datasets}
