% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_data_array.R
\name{env_data_array}
\alias{env_data_array}
\title{envDataArray
Get an array of environmental data from presence-absence points.}
\usage{
env_data_array(env_data, occ = NULL)
}
\arguments{
\item{env_data}{List of environmental variables time series stack.}

\item{occ}{Occurrence data frame. Should contain longitude and latitude
columns.}
}
\value{
An array of M points times N time steps times P environmental
variables
}
\description{
envDataArray
Get an array of environmental data from presence-absence points.
}
\examples{
occ <- mus_virtualis[1:5, ]
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
env_data <- list(bio1 = bio1_ts, bio12 = bio12_ts)
env_data_array(env_data, occ)
}
