% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envdat_ex.R
\docType{data}
\name{envdat_ex}
\alias{envdat_ex}
\title{Environmental array}
\format{
An array with with dimensions 2,39 and 200:
}
\usage{
envdat_ex
}
\description{
An array of a time series of 30 years of bio1 and bio12 bioclimatic
variables
}
\details{
the dimenssion 2 correspond tot he bioclimatic variables. The dimension
39 correspond to each year of the time serires and the dimenssion 2000
correspond to the data points.
}
\keyword{datasets}
