% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envDataArray.R
\name{envDataArray}
\alias{envDataArray}
\title{envDataArray
Get an array of environmental data from presence-absence points.}
\usage{
envDataArray(envData, occ = NULL)
}
\arguments{
\item{envData}{List of environmental variables time series stack.}

\item{occ}{Occurence data frame. Should contain longitude and latitude columns}
}
\value{
An array of M points times N time steps times P environmental variables
}
\description{
envDataArray
Get an array of environmental data from presence-absence points.
}
\examples{
occ <- mus_virtualis[1:5, ]
bio1_ts <- terra::unwrap(cmcc_cm_bio1)
bio12_ts <- terra::unwrap(cmcc_cm_bio12)
envData <- list(bio1 = bio1_ts, bio12 = bio12_ts)
envDataArray(envData, occ)
}
