% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit_gen.R
\name{expit_gen}
\alias{expit_gen}
\title{Generalized expit function}
\usage{
expit_gen(x, L, U, x0)
}
\arguments{
\item{x}{A real number to ma}

\item{L}{Lower bound}

\item{U}{Upper bound}

\item{x0}{Middle point}
}
\value{
A rescaled valued between L and U with the middle point x0
}
\description{
Generalized expit function
}
\examples{
expit_gen(3, -pi, pi, 0)
}
