% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_range_df.R
\name{get_range_df}
\alias{get_range_df}
\title{Get range data frame. Creates a unfilled data frame
to store the parameter ranges given a number of environmental variables}
\usage{
get_range_df(envdat, quant_vec = c(0.1, 0.5, 0.9))
}
\arguments{
\item{envdat}{An environmental array with dimenssions given by 
environmental variables, time serires steps and locations.
Should provide an environmental array where the detections are possitive}

\item{quant_vec}{A vector of quantiles to limit the range of the
parameters given the occurrence in the environmental time serires.
Default is set to c(0.1, 0.5, 0.9) to center in the median value}
}
\value{
A data.frame with NA three columns, lower bound, center and
upper bound of the range of the parameters.
}
\description{
Get range data frame. Creates a unfilled data frame
to store the parameter ranges given a number of environmental variables
}
