% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob_detect.R
\name{log_prob_detect}
\alias{log_prob_detect}
\title{Probability of Detection}
\usage{
log_prob_detect(
  env_dat,
  mu,
  sigl,
  sigr,
  ctil,
  pd,
  o_mat,
  return_prob = FALSE,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{env_dat}{The environmental data array}

\item{mu}{A vector of mu}

\item{sigl}{A vector of sigl}

\item{sigr}{A vector of sigr}

\item{ctil}{C tilde parameter}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{o_mat}{An orthogonal matrix}

\item{return_prob}{Logical (default FALSE) Flag to return
probability of detection instead the log of probability of detection}

\item{num_threads}{The number of threads for parallel computing}
}
\value{
A numeric value of the the probability (or log prob) of detection
}
\description{
Probability of Detection
}
\examples{
mu <- c(11.433373, 5.046939)
sigl <- c(1.036834, 1.556083)
sigr <- c(1.538972, 1.458738)
ctil <- -2
pd <- 0.9
o_mat <- matrix(c(-0.4443546, 0.8958510, -0.8958510, -0.4443546), ncol = 2)
M <- log_prob_detect(envdat_ex, mu, sigl, sigl, ctil, pd, o_mat)
}
