% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_biol.R
\name{loglik_biol}
\alias{loglik_biol}
\title{Logarithm of the likelihood of detection of a species}
\usage{
loglik_biol(
  env_dat,
  occ,
  mu,
  sigl,
  sigr,
  ctil,
  pd,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads(),
  sum_log_p = TRUE,
  return_prob = FALSE
)
}
\arguments{
\item{env_dat}{The environmental data array. This is a three dimensional
array of dimensions (number of locations of species detection or
pseudo-absence) by (time series length) by (number of environmental variables
considered).}

\item{occ}{Presence absence binary vector. Same length as dimension 1 of
env_dat.}

\item{mu}{A vector of mu parameters. Unconstrained real numbers. Same length
as dimension 3 of env_dat.}

\item{sigl}{A vector of sigl parameters}

\item{sigr}{A vector of sigr parameters}

\item{ctil}{C tilde parameter}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{o_mat}{An orthogonal matrix}

\item{num_threads}{The number of threads for parallel computing}

\item{sum_log_p}{Logical (default TRUE) returns the sum of log of probability
detection}

\item{return_prob}{Logical (default FALSE) returns the probability of
detections instead the log of probabilities}
}
\value{
The value of the logarithm of the likelihood function in the
provided parameters for a given set of environmetnal conditions an presence
absence locations
}
\description{
Logarithm of the likelihood of detection of a species
}
\examples{
o_mat <- param_list_example$o_mat
mu <- param_list_example$mu
sigl <- param_list_example$sigl
sigr <- param_list_example$sigr
pd <- param_list_example$pd
ctil <- param_list_example$ctil
env_dat <- envdat_ex
occ <- occExample

ll <- loglik_biol(
  env_dat,
  occ,
  mu = mu,
  sigl = sigr,
  sigr = sigl,
  ctil = ctil,
  pd = pd,
  o_mat = o_mat
)
ll
}
