% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_math.R
\name{loglik_math}
\alias{loglik_math}
\title{log likelihood orthogonal}
\usage{
loglik_math(
  param_vector,
  env_dat,
  occ,
  opt = NULL,
  num_threads = RcppParallel::defaultNumThreads(),
  negative = TRUE
)
}
\arguments{
\item{param_vector}{A vector with parameters to test in a log-likelihood
function of the species distribution model.}

\item{env_dat}{An array with environmental data, includes the time series
in each location for each environmental variable}

\item{occ}{A presence absence vector}

\item{opt}{An optional vector to mute a specific parameter to evaluate. It is
further used to profile the optimization}

\item{num_threads}{The number of threads to run the main function}

\item{negative}{Logical. If true returns the negative of the log-likelihood
function. Flag should be set when optimize.}
}
\value{
A value with the log of the likelihood for a given set of
parameters, environmental variables and presence-absences occurrences.
}
\description{
log likelihood orthogonal
}
\examples{
env_dat <- envdat_ex
occ <- occExample
params <- param_table_example[5, ]
param_vector <- params
loglik_math(params, env_dat, occ)

opt_a <- c(
  mu1 = NA_real_,
  mu2 = 0,
  sigl1 = NA_real_,
  sigl2 = NA_real_,
  sigr1 = NA_real_,
  sigr2 = NA_real_,
  ctil = NA_real_,
  pd = NA_real_,
  opar_1 = NA_real_
)
loglik_math(params, env_dat, occ, opt = opt_a)
}
