% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_orthog_nd_unconstr.R
\name{loglik_orthog_nd_unconstr}
\alias{loglik_orthog_nd_unconstr}
\title{log likelihood orthogonal}
\usage{
loglik_orthog_nd_unconstr(
  param_vector,
  env_dat,
  occ,
  opt = NULL,
  negative = TRUE,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{param_vector}{A vector with parameters to test in a log-likelihood
function of the species distribution model.}

\item{env_dat}{An array with environmental data, includes the time series
in each location for each environmental variable}

\item{occ}{A presence absence vector}

\item{opt}{An optional vector to mute a specific parameter to evaluate. It is
further used to profile the optimization}

\item{negative}{Logical. If true returns the negative of the log-likelihood
function. Flag should be set when optimize.}

\item{num_threads}{The number of threads to run the main function}
}
\value{
A value with the log of the likelihood for a given set of
parameters, environmental variables and presence-absences ocurrences.
}
\description{
log likelihood orthogonal
}
\examples{
env_dat <- envdat_ex
occ <- occExample
params <- paramTableExample[5, ]
param_vector <- params
loglik_orthog_nd_unconstr(params, env_dat, occ)
}
