% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logprobdetect.R
\name{logprobdetect}
\alias{logprobdetect}
\title{Title}
\usage{
logprobdetect(
  envdat,
  mu,
  sigLtil,
  sigRtil,
  ctil,
  pd,
  O,
  return_prob = FALSE,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{envdat}{The environmental data array}

\item{mu}{A vector of mu}

\item{sigLtil}{A vector of sigl}

\item{sigRtil}{A vector of sigr}

\item{ctil}{C tilde parameter}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{O}{An orthogonal matrix}

\item{return_prob}{Logical (default FALSE) Flag to return
probability of detection instead the log of probability of detection}

\item{num_threads}{The number of threads for parallel computing}
}
\value{
A numeric value of the the probability (or log prob) of detection
}
\description{
Title
}
\examples{
mu = c(11.433373, 5.046939)
sigLtil = c(1.036834, 1.556083)
sigRtil = c(1.538972, 1.458738)
ctil = -2
pd = 0.9
O = matrix(c(-0.4443546, 0.8958510, -0.8958510, -0.4443546), ncol = 2)
M <- logprobdetect(envdat_ex, mu, sigRtil, sigRtil,ctil, pd, O)
}
