% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logprobdetect_r.R
\name{logprobdetect_r}
\alias{logprobdetect_r}
\title{Title}
\usage{
logprobdetect_r(O, mu, sigLtil, sigRtil, pd, ctil, envdat)
}
\arguments{
\item{O}{An orthogonal matrix}

\item{mu}{A vector of mu}

\item{sigLtil}{A vector of sigl}

\item{sigRtil}{A vector of sigr}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{ctil}{C tilde parameter}

\item{envdat}{The environmental data array}
}
\value{
A numeric value of the log of the probability of detection
}
\description{
Title
}
\examples{
O = matrix(c(-0.4443546, 0.8958510, -0.8958510, -0.4443546), ncol = 2)
mu = c(11.433373, 5.046939)
sigLtil = c(1.036834, 1.556083)
sigRtil = c(1.538972, 1.458738)
pd = 0.9
ctil = -1
M <- logprobdetect_r(O, mu, sigLtil, sigRtil, pd, ctil, envdat_ex)
}
