% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_canonical.R
\name{make_cannonical}
\alias{make_cannonical}
\title{Title}
\usage{
make_cannonical(bparms)
}
\arguments{
\item{bparms}{A list with parameters
to adjust to a canonical form}
}
\value{
A list  with parameters in canonical form
}
\description{
Title
}
\examples{

paramTableExample[1, ] |> 
math_to_bio() |>
 make_cannonical()
}
