% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_canonical.R
\name{make_canonical}
\alias{make_canonical}
\title{This is used to calculated horizontal distance
between parameters}
\usage{
make_canonical(parms_bio)
}
\arguments{
\item{parms_bio}{A list with parameters to adjust to a canonical form}
}
\value{
A list  with parameters in canonical form
}
\description{
This is used to calculated horizontal distance
between parameters
}
\examples{

param_table_example[1, ] |>
  math_to_bio() |>
  make_canonical()
}
