% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_to_bio.R
\name{math_to_bio}
\alias{math_to_bio}
\title{Convert a unconstrained vector of parameters in a list of parameters
to fit the log likelihood function.}
\usage{
math_to_bio(param_vector)
}
\arguments{
\item{param_vector}{A vector of unconstrained parameters to create a list for
the likelihood function}
}
\value{
A list of parameters to pass to the log likelihood function.
It contains o_par, mu, sigl, sigr, ctil, pd
}
\description{
Convert a unconstrained vector of parameters in a list of parameters
to fit the log likelihood function.
}
\examples{
math_to_bio(param_table_example[1, ])
}
