% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numparms.R
\name{numparms}
\alias{numparms}
\title{Get the number of parameters of the model given the number
of environmental time serires}
\usage{
numparms(p)
}
\arguments{
\item{p}{An integer with the number of environmental variables}
}
\value{
An integer with the number of parameters
}
\description{
Get the number of parameters of the model given the number
of environmental time serires
}
\examples{
numparms(2)
}
