% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occExample.R
\docType{data}
\name{occExample}
\alias{occExample}
\title{Ocurrence example}
\format{
An binary (ones and zeros) vector of ocurrence of a virtual species.
}
\usage{
occExample
}
\description{
A vector of 2000 ocurrencies of a Virtual species
}
\details{
It helpts to test the log likelihood function
}
\keyword{datasets}
