% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramTableExample.R
\docType{data}
\name{paramTableExample}
\alias{paramTableExample}
\title{Environmental parameter table}
\format{
A data.frame with with dimensions 9 columns and 100 rows:
The parameters are ctil, O1, mu1, mu2, sigLtil1, sigLtil2, sigRtil1, sigRtil2, pd
}
\usage{
paramTableExample
}
\description{
An data frame with a set of 100 parameters to test the
optimization of the likelihood function
}
\details{
This set of parameters correspond to the case of a two environmental
time series fit
}
\keyword{datasets}
