% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_optim_par.R
\name{profile_optim_par}
\alias{profile_optim_par}
\title{Title}
\usage{
profile_optim_par(optim_vec, envdat, pa, stepsize = 0.001, parallel = FALSE)
}
\arguments{
\item{optim_vec}{Vector of optim parameters
fitted with optim_ll function}

\item{envdat}{Environmental dataset}

\item{pa}{presence absence vector}

\item{stepsize}{The stepsize of the profiler}

\item{parallel}{Run the profiles in parallel}
}
\value{
A data.frame with the profile for each
parameter
optimVector <- mll_df_example[1, 1:9] |> unlist()
profile_optim_par(optimVector, stepsize = 1)
}
\description{
Title
}
