% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_virtual_example.R
\docType{data}
\name{sp_virtual_example}
\alias{sp_virtual_example}
\title{sp_virtual_example}
\format{
A data frame with 60 rows and 4 variables:
\describe{
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{presence}}{double COLUMN_DESCRIPTION}
}
}
\usage{
sp_virtual_example
}
\description{
An example of a virtual species
}
\details{
DETAILS
}
\keyword{datasets}
