% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_parms.R
\name{start_parms}
\alias{start_parms}
\title{startparms Starting parameters for the optimization.
Given a set of ranges of values of environmental variables where the
presence of a species occurs,generates a Latin hypercube design
for the parameters based on the Sobol' low-discrepancy sequence.}
\usage{
start_parms(env_dat, quant_vec = c(0.1, 0.5, 0.9), num_starts = 100)
}
\arguments{
\item{env_dat}{The environmental array given the observed occurrences}

\item{quant_vec}{A vector of quantiles. Are the bounds of the range to create
the hypercube}

\item{num_starts}{The number of samples of the hypercube}
}
\value{
A data frame with samples for each parameter to optimize
}
\description{
startparms Starting parameters for the optimization.
Given a set of ranges of values of environmental variables where the
presence of a species occurs,generates a Latin hypercube design
for the parameters based on the Sobol' low-discrepancy sequence.
}
\examples{
envdat_ex_occ <- envdat_ex[, , occExample == 1]
start_parms(envdat_ex_occ)
}
