% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startparms.R
\name{startparms}
\alias{startparms}
\title{startparms Starting parameters for the optimization.
Given a set of ranges of values of environmental variables where the 
presence of a species occurs,generates a Latin hypercube design
for the parameters based on the Sobol' low-discrepancy sequence.}
\usage{
startparms(envdat, quant_vec = c(0.1, 0.5, 0.9), numstarts = 100)
}
\arguments{
\item{envdat}{The environmental array given the observed ocurrences}

\item{quant_vec}{A vector of quantiles. Are the bounds of the range to create
the hypercube}

\item{numstarts}{The number of samples of the hypercube}
}
\value{
A data frame with samples for each parameter to optimize
}
\description{
startparms Starting parameters for the optimization.
Given a set of ranges of values of environmental variables where the 
presence of a species occurs,generates a Latin hypercube design
for the parameters based on the Sobol' low-discrepancy sequence.
}
\examples{
envdat_ex_occ <- envdat_ex[ , , occExample == 1] 
startparms(envdat_ex_occ)
}
