% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdmMle-package.R
\docType{package}
\name{xsdmMle-package}
\alias{xsdmMle}
\alias{xsdmMle-package}
\title{xsdmMle: Concise Summary of What the Package Does}
\description{
More about what it does (maybe more than one line).
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

}
