% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_neg_ltsgr_cpp.R
\name{like_neg_ltsgr_cpp}
\alias{like_neg_ltsgr_cpp}
\title{uasym}
\usage{
like_neg_ltsgr_cpp(
  envdat,
  mu,
  sigLtil,
  sigRtil,
  O,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{envdat}{The environmental data array}

\item{mu}{A vector of mu}

\item{sigLtil}{A vector of sigl}

\item{sigRtil}{A vector of sigr}

\item{O}{An orthogonal matrix}

\item{num_threads}{The number of threads for parallel computing}
}
\value{
A numeric value of the log of the likelihood function
}
\description{
uasym
}
\examples{
O = matrix(c(-0.4443546, 0.8958510, -0.8958510, -0.4443546), ncol = 2)
mu = c(11.433373, 5.046939)
sigLtil = c(1.036834, 1.556083)
sigRtil = c(1.538972, 1.458738)
M <- like_neg_ltsgr_cpp(envdat_ex, mu, sigLtil, sigLtil, O)
}
