% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_neg_ltsgr_cpp.R
\name{like_neg_ltsgr_cpp}
\alias{like_neg_ltsgr_cpp}
\title{The negative of the (like) long term stochastich growth function
implemented in c++.}
\usage{
like_neg_ltsgr_cpp(
  env_dat,
  mu,
  sigl,
  sigr,
  o_mat,
  n_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{env_dat}{The environmental data array}

\item{mu}{A vector of mu}

\item{sigl}{A vector of sigl}

\item{sigr}{A vector of sigr}

\item{o_mat}{An orthogonal matrix}

\item{n_threads}{The number of threads for parallel computing}
}
\value{
A numeric value of the log of the likelihood function
}
\description{
The negative of the (like) long term stochastich growth function
implemented in c++.
}
\examples{
o_mat <- matrix(c(-0.4443546, 0.8958510, -0.8958510, -0.4443546), ncol = 2)
mu <- c(11.433373, 5.046939)
sigl <- c(1.036834, 1.556083)
sigr <- c(1.538972, 1.458738)
M <- like_neg_ltsgr_cpp(env_dat = envdat_ex,
mu = mu,
sigl = sigl,
sigr = sigl,
o_mat = o_mat)
}
