% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_neg_ltsgr_r.R
\name{like_neg_ltsgr_r}
\alias{like_neg_ltsgr_r}
\title{Log-Likelihood Function for Asymmetric Environmental Long Stochastic Growth}
\usage{
like_neg_ltsgr_r(envdat, mu, sigLtil, sigRtil, O)
}
\arguments{
\item{envdat}{The environmental data array}

\item{mu}{A vector of mu}

\item{sigLtil}{A vector of sigl}

\item{sigRtil}{A vector of sigr}

\item{O}{An orthogonal matrix}
}
\value{
A numeric value of the log of the likelihood function
}
\description{
Log-Likelihood Function for Asymmetric Environmental Long Stochastic Growth
}
