% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1mexp.R
\name{log1mexp}
\alias{log1mexp}
\title{Compute f(a) = log(1 +/- exp(-a)) Numerically Optimally.
Original code in copula R package.}
\usage{
log1mexp(a, cutoff = log(2))
}
\arguments{
\item{a}{numeric vector of positive values}

\item{cutoff}{positive number; log(2) is "optimal", but the exact value is
unimportant, and anything in [0.5, 1] is fine.}
}
\value{
f(a) == log(1 - exp(-a)) == log1p(-exp(-a)) == log(-expm1(-a))
or g(x) == log(1 + exp(x)) == log1p(exp(x)) computed accurately and quickly
}
\description{
Compute f(a) = log(1 +/- exp(-a)) Numerically Optimally.
Original code in copula R package.
}
\examples{
a <- 2^seq(-58, 10, length = 256)
f_expr <- expression(
  log(1 - exp(-a)),
  log(-expm1(-a)),
  log1p(-exp(-a)),
  log1mexp(a)
)
names(f_expr) <- c("DEF", "expm1", "log1p", "F")
str(fa <- do.call(cbind, as.list(f_expr)))
head(fa) # expm1() works here
tail(fa) # log1p() works here
}
