% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log1pexp.R
\name{log1pexp}
\alias{log1pexp}
\title{Compute f(a) = log(1 +/- exp(-a)) Numerically Optimally.
Original function by Martin Mächler (2012) in copula R pacakge}
\usage{
log1pexp(x, c0 = -37, c1 = 18, c2 = 33.3)
}
\arguments{
\item{x}{Numeric vector}

\item{c0}{Cutoffs for log1pexp}

\item{c1}{Cutoffs for log1pexp}

\item{c2}{Cutoffs for log1pexp}
}
\value{
f(a) == log(1 - exp(-a)) == log1p(-exp(-a)) == log(-expm1(-a))
or g(x) == log(1 + exp(x)) == log1p(exp(x))
computed accurately and quickly
}
\description{
Compute f(a) = log(1 +/- exp(-a)) Numerically Optimally.
Original function by Martin Mächler (2012) in copula R pacakge
}
\examples{
x <- seq(700, 720, by = 2)
cbind(x, log1p(exp(x)), log1pexp(x))
}
