% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_orthog_nd.R
\name{loglik_orthog_nd}
\alias{loglik_orthog_nd}
\title{Logarithm of the likelihood oof detection of a species}
\usage{
loglik_orthog_nd(
  O,
  mu,
  sigLtil,
  sigRtil,
  pd,
  ctil,
  envdat,
  pa,
  num_threads = RcppParallel::defaultNumThreads()
)
}
\arguments{
\item{O}{An orthogonal matrix}

\item{mu}{A vector of mu}

\item{sigLtil}{A vector of sigl}

\item{sigRtil}{A vector of sigr}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{ctil}{C tilde parameter}

\item{envdat}{The environmental data array}

\item{pa}{Presence absence binary vector}

\item{num_threads}{The number of threads for parallel computing}
}
\value{
The value of the logarithm of the likelihood function in the 
provided parameters for a given set of environmetnal conditions an presence
absence locations
}
\description{
Logarithm of the likelihood oof detection of a species
}
\examples{
O <-  paramListExample$O
mu <- paramListExample$mu
sigLtil <- paramListExample$sigLtil
sigRtil <- paramListExample$sigRtil
pd <- paramListExample$pd
ctil <- paramListExample$ctil
envdat <- envdat_ex
pa <- occExample

ll <- loglik_orthog_nd(O = O, mu = mu, sigLtil = sigLtil,sigRtil =  sigLtil, pd = pd, ctil = ctil, 
envdat_ex, pa )
}
