% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_orthog_nd.R
\name{loglik_orthog_nd}
\alias{loglik_orthog_nd}
\title{Logarithm of the likelihood oof detection of a species}
\usage{
loglik_orthog_nd(
  env_dat,
  occ,
  mu,
  sigl,
  sigr,
  ctil,
  pd,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads(),
  sum_logp = TRUE,
  return_prob = FALSE
)
}
\arguments{
\item{env_dat}{The environmental data array}

\item{occ}{Presence absence binary vector}

\item{mu}{A vector of mu parameters}

\item{sigl}{A vector of sigl parameters}

\item{sigr}{A vector of sigr parameters}

\item{ctil}{C tilde parameter}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{o_mat}{An orthogonal matrix}

\item{num_threads}{The number of threads for parallel computing}

\item{sum_logp}{Logical (default TRUE) returns the sum of log of probability
detection}

\item{return_prob}{Logical (default FALSE) returns the probability of
detections instead the log of probabilities}
}
\value{
The value of the logarithm of the likelihood function in the
provided parameters for a given set of environmetnal conditions an presence
absence locations
}
\description{
Logarithm of the likelihood oof detection of a species
}
\examples{
o_mat <- param_list_example$o_mat
mu <- param_list_example$mu
sigl <- param_list_example$sigl
sigr <- param_list_example$sigr
pd <- param_list_example$pd
ctil <- param_list_example$ctil
env_dat <- envdat_ex
occ <- occExample

ll <- loglik_orthog_nd(
  env_dat,
  occ,
  mu = mu,
  sigl = sigr,
  sigr = sigl,
  ctil = ctil,
  pd = pd,
  o_mat = o_mat
)
ll
}
