% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_orthog_nd.R
\name{loglik_orthog_nd}
\alias{loglik_orthog_nd}
\title{Logarithm of the likelihood oof detection of a species}
\usage{
loglik_orthog_nd(
  envdat,
  occ,
  mu,
  sigl,
  sigr,
  ctil,
  pd,
  o_mat,
  num_threads = RcppParallel::defaultNumThreads(),
  sum_logp = TRUE,
  return_prob = FALSE
)
}
\arguments{
\item{envdat}{The environmental data array}

\item{occ}{Presence absence binary vector}

\item{mu}{A vector of mu}

\item{sigl}{A vector of sigl}

\item{sigr}{A vector of sigr}

\item{ctil}{C tilde parameter}

\item{pd}{Probability of detection. Parameter between 0 and 1}

\item{o_mat}{An orthogonal matrix}

\item{num_threads}{The number of threads for parallel computing}

\item{sum_logp}{Logical (default TRUE) returns the sum of log of probability
detection}

\item{return_prob}{Logical (default FALSE) returns the probability of
detections instead the log of probabilities}
}
\value{
The value of the logarithm of the likelihood function in the
provided parameters for a given set of environmetnal conditions an presence
absence locations
}
\description{
Logarithm of the likelihood oof detection of a species
}
\examples{
o_mat <- paramListExample$O
mu <- paramListExample$mu
sigl <- paramListExample$sigLtil
sigr <- paramListExample$sigRtil
pd <- paramListExample$pd
ctil <- paramListExample$ctil
envdat <- envdat_ex
occ <- occExample

ll <- loglik_orthog_nd(envdat_ex, occ,
  mu = mu,
  sigl = sigl,
  sigr = sigl,
  ctil = ctil, pd = pd, o_mat = o_mat
)
ll
}
