% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_mll.R
\name{optim_mll}
\alias{optim_mll}
\title{Optimization of the log-likelihood function}
\usage{
optim_mll(envdat, occ, parallel = FALSE, numstarts = 100)
}
\arguments{
\item{envdat}{Environmental time series array}

\item{occ}{Presence absence vector}

\item{parallel}{Parallelization strategy}

\item{numstarts}{Number of samples to start the optimizations.}
}
\description{
Optimization of the log-likelihood function
}
\examples{
optim_df <- optim_mll(envdat_ex[, , 1:10], occExample[1:10], numstarts = 10)

}
