% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\title{Virtual species}
\usage{
vsp(env_data, param_list)
}
\arguments{
\item{env_data}{A list of time series rasters of bioclimatic variables}

\item{param_list}{A list of parameters}
}
\value{
A raster with a virtual species probability of detection
}
\description{
Virtual species
}
\examples{
# example code
paramListExample
xsdmMle::envdat_ex
vsp(envdat_ex, param_list_example)
}
