% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsp.R
\name{vsp}
\alias{vsp}
\title{Virtual species}
\usage{
vsp(envData, param_list)
}
\arguments{
\item{envData}{A list of time serires rasters of bioclimatic variables}

\item{param_list}{A list of parameters}
}
\value{
A raster with a virtual species probability of detection
}
\description{
Virtual species
}
