% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsdmMle-package.R
\docType{package}
\name{xsdmMle-package}
\alias{xsdmMle}
\alias{xsdmMle-package}
\title{xsdmMle: Demographic Approach to Species Distribution Model}
\description{
Integrates concepts of stochastic demography into species distribution modelling. The main approach is maximize the likelihood function with environmental information from the presence/absence. This inference can be used to project species geographic range. Data requirements include species detection records and spatiotemporal environmental data.
}
\author{
\strong{Maintainer}: Angel Luis Robles Fernandez \email{a.l.robles.fernandez@gmail.com} (\href{https://orcid.org/0000-0002-4674-4270}{ORCID})

Authors:
\itemize{
  \item Emilio Berti \email{emilio.berti@idiv.de} (\href{https://orcid.org/0000-0001-9286-011X}{ORCID})
  \item Daniel C. Reuman \email{reuman@ku.edu} (\href{https://orcid.org/0000-0002-1407-8947}{ORCID})
}

}
